8367014: Change prefix for platform-dependent AtomicAccess files
