+++ src/hotspot/os_cpu/aix_ppc/atomicAccess_aix_ppc.hpp
@@ -23,17 +23,17 @@
  *
  */
 
-#ifndef OS_CPU_LINUX_PPC_ATOMIC_LINUX_PPC_HPP
-#define OS_CPU_LINUX_PPC_ATOMIC_LINUX_PPC_HPP
+#ifndef OS_CPU_AIX_PPC_ATOMICACCESS_AIX_PPC_HPP
+#define OS_CPU_AIX_PPC_ATOMICACCESS_AIX_PPC_HPP
 
 #ifndef PPC64
 #error "Atomic currently only implemented for PPC64"
 #endif
 
-#include "orderAccess_linux_ppc.hpp"
+#include "orderAccess_aix_ppc.hpp"
 #include "utilities/debug.hpp"
 
-// Implementation of class atomic
+// Implementation of class AtomicAccess
 
 //
 // machine barrier instructions:
@@ -246,30 +246,53 @@ inline T AtomicAccess::PlatformCmpxchg<1>::operator()(T volatile* dest,
   // specified otherwise (see atomicAccess.hpp).
 
   // Using 32 bit internally.
-  unsigned int old_value, loaded_value;
+  volatile int *dest_base = (volatile int*)((uintptr_t)dest & ~3);
+
+#ifdef VM_LITTLE_ENDIAN
+  const unsigned int shift_amount        = ((uintptr_t)dest & 3) * 8;
+#else
+  const unsigned int shift_amount        = ((~(uintptr_t)dest) & 3) * 8;
+#endif
+  const unsigned int masked_compare_val  = ((unsigned int)(unsigned char)compare_value),
+                     masked_exchange_val = ((unsigned int)(unsigned char)exchange_value),
+                     xor_value           = (masked_compare_val ^ masked_exchange_val) << shift_amount;
+
+  unsigned int old_value, value32;
+
   pre_membar(order);
 
   __asm__ __volatile__ (
+    /* simple guard */
+    "   lbz     %[old_value], 0(%[dest])                  \n"
+    "   cmpw    %[masked_compare_val], %[old_value]       \n"
+    "   bne-    2f                                        \n"
     /* atomic loop */
     "1:                                                   \n"
-    "   lbarx   %[old_value], 0, %[dest]                  \n"
+    "   lwarx   %[value32], 0, %[dest_base]               \n"
     /* extract byte and compare */
-    "   cmpw    %[compare_value], %[old_value]            \n"
+    "   srd     %[old_value], %[value32], %[shift_amount] \n"
+    "   clrldi  %[old_value], %[old_value], 56            \n"
+    "   cmpw    %[masked_compare_val], %[old_value]       \n"
     "   bne-    2f                                        \n"
     /* replace byte and try to store */
-    "   stbcx.  %[exchange_value], 0, %[dest]             \n"
+    "   xor     %[value32], %[xor_value], %[value32]      \n"
+    "   stwcx.  %[value32], 0, %[dest_base]               \n"
     "   bne-    1b                                        \n"
     /* exit */
     "2:                                                   \n"
     /* out */
     : [old_value]           "=&r"   (old_value),
-      [loaded_value]        "=&r"   (loaded_value),
-                            "=m"    (*dest)
+      [value32]             "=&r"   (value32),
+                            "=m"    (*dest),
+                            "=m"    (*dest_base)
     /* in */
-    : [dest]            "b"     (dest),
-      [compare_value]   "r"     (compare_value),
-      [exchange_value]  "r"     (exchange_value),
-                        "m"     (*dest)
+    : [dest]                "b"     (dest),
+      [dest_base]           "b"     (dest_base),
+      [shift_amount]        "r"     (shift_amount),
+      [masked_compare_val]  "r"     (masked_compare_val),
+      [xor_value]           "r"     (xor_value),
+                            "m"     (*dest),
+                            "m"     (*dest_base)
     /* clobber */
     : "cc",
       "memory"
@@ -381,8 +404,7 @@ inline T AtomicAccess::PlatformCmpxchg<8>::operator()(T volatile* dest,
 }
 
 template<size_t byte_size>
-struct AtomicAccess::PlatformOrderedLoad<byte_size, X_ACQUIRE>
-{
+struct AtomicAccess::PlatformOrderedLoad<byte_size, X_ACQUIRE> {
   template <typename T>
   T operator()(const volatile T* p) const {
     T t = AtomicAccess::load(p);
@@ -392,4 +414,4 @@ struct AtomicAccess::PlatformOrderedLoad<byte_size, X_ACQUIRE>
   }
 };
 
-#endif // OS_CPU_LINUX_PPC_ATOMIC_LINUX_PPC_HPP
+#endif // OS_CPU_AIX_PPC_ATOMICACCESS_AIX_PPC_HPP
