+++ src/hotspot/os_cpu/bsd_x86/atomicAccess_bsd_x86.hpp
@@ -22,15 +22,15 @@
  *
  */
 
-#ifndef OS_CPU_LINUX_X86_ATOMIC_LINUX_X86_HPP
-#define OS_CPU_LINUX_X86_ATOMIC_LINUX_X86_HPP
+#ifndef OS_CPU_BSD_X86_ATOMICACCESS_BSD_X86_HPP
+#define OS_CPU_BSD_X86_ATOMICACCESS_BSD_X86_HPP
 
-// Implementation of class atomic
+// Implementation of class AtomicAccess
 
 template<size_t byte_size>
 struct AtomicAccess::PlatformAdd {
   template<typename D, typename I>
-  D fetch_then_add(D volatile* dest, I add_value, atomic_memory_order order) const;
+  D fetch_then_add(D volatile* dest, I add_value, atomic_memory_order /* order */) const;
 
   template<typename D, typename I>
   D add_then_fetch(D volatile* dest, I add_value, atomic_memory_order order) const {
@@ -41,7 +41,7 @@ struct AtomicAccess::PlatformAdd {
 template<>
 template<typename D, typename I>
 inline D AtomicAccess::PlatformAdd<4>::fetch_then_add(D volatile* dest, I add_value,
-                                                      atomic_memory_order order) const {
+                                                      atomic_memory_order /* order */) const {
   STATIC_ASSERT(4 == sizeof(I));
   STATIC_ASSERT(4 == sizeof(D));
   D old_value;
@@ -56,7 +56,7 @@ template<>
 template<typename T>
 inline T AtomicAccess::PlatformXchg<4>::operator()(T volatile* dest,
                                                    T exchange_value,
-                                                   atomic_memory_order order) const {
+                                                   atomic_memory_order /* order */) const {
   STATIC_ASSERT(4 == sizeof(T));
   __asm__ volatile (  "xchgl (%2),%0"
                     : "=r" (exchange_value)
@@ -72,7 +72,7 @@ inline T AtomicAccess::PlatformCmpxchg<1>::operator()(T volatile* dest,
                                                       T exchange_value,
                                                       atomic_memory_order /* order */) const {
   STATIC_ASSERT(1 == sizeof(T));
-  __asm__ volatile ("lock cmpxchgb %1,(%3)"
+  __asm__ volatile (  "lock cmpxchgb %1,(%3)"
                     : "=a" (exchange_value)
                     : "q" (exchange_value), "a" (compare_value), "r" (dest)
                     : "cc", "memory");
@@ -86,7 +86,7 @@ inline T AtomicAccess::PlatformCmpxchg<4>::operator()(T volatile* dest,
                                                       T exchange_value,
                                                       atomic_memory_order /* order */) const {
   STATIC_ASSERT(4 == sizeof(T));
-  __asm__ volatile ("lock cmpxchgl %1,(%3)"
+  __asm__ volatile (  "lock cmpxchgl %1,(%3)"
                     : "=a" (exchange_value)
                     : "r" (exchange_value), "a" (compare_value), "r" (dest)
                     : "cc", "memory");
@@ -94,15 +94,14 @@ inline T AtomicAccess::PlatformCmpxchg<4>::operator()(T volatile* dest,
 }
 
 #ifdef AMD64
-
 template<>
 template<typename D, typename I>
 inline D AtomicAccess::PlatformAdd<8>::fetch_then_add(D volatile* dest, I add_value,
-                                                      atomic_memory_order order) const {
+                                                      atomic_memory_order /* order */) const {
   STATIC_ASSERT(8 == sizeof(I));
   STATIC_ASSERT(8 == sizeof(D));
   D old_value;
-  __asm__ __volatile__ ("lock xaddq %0,(%2)"
+  __asm__ __volatile__ (  "lock xaddq %0,(%2)"
                         : "=r" (old_value)
                         : "0" (add_value), "r" (dest)
                         : "cc", "memory");
@@ -111,8 +110,9 @@ inline D AtomicAccess::PlatformAdd<8>::fetch_then_add(D volatile* dest, I add_va
 
 template<>
 template<typename T>
-inline T AtomicAccess::PlatformXchg<8>::operator()(T volatile* dest, T exchange_value,
-                                                   atomic_memory_order order) const {
+inline T AtomicAccess::PlatformXchg<8>::operator()(T volatile* dest,
+                                                   T exchange_value,
+                                                   atomic_memory_order /* order */) const {
   STATIC_ASSERT(8 == sizeof(T));
   __asm__ __volatile__ ("xchgq (%2),%0"
                         : "=r" (exchange_value)
@@ -128,7 +128,7 @@ inline T AtomicAccess::PlatformCmpxchg<8>::operator()(T volatile* dest,
                                                       T exchange_value,
                                                       atomic_memory_order /* order */) const {
   STATIC_ASSERT(8 == sizeof(T));
-  __asm__ __volatile__ ("lock cmpxchgq %1,(%3)"
+  __asm__ __volatile__ (  "lock cmpxchgq %1,(%3)"
                         : "=a" (exchange_value)
                         : "r" (exchange_value), "a" (compare_value), "r" (dest)
                         : "cc", "memory");
@@ -138,7 +138,7 @@ inline T AtomicAccess::PlatformCmpxchg<8>::operator()(T volatile* dest,
 #else // !AMD64
 
 extern "C" {
-  // defined in linux_x86.s
+  // defined in bsd_x86.s
   int64_t _Atomic_cmpxchg_long(int64_t, volatile int64_t*, int64_t);
   void _Atomic_move_long(const volatile int64_t* src, volatile int64_t* dst);
 }
@@ -148,7 +148,7 @@ template<typename T>
 inline T AtomicAccess::PlatformCmpxchg<8>::operator()(T volatile* dest,
                                                       T compare_value,
                                                       T exchange_value,
-                                                      atomic_memory_order order) const {
+                                                      atomic_memory_order /* order */) const {
   STATIC_ASSERT(8 == sizeof(T));
   return cmpxchg_using_helper<int64_t>(_Atomic_cmpxchg_long, dest, compare_value, exchange_value);
 }
@@ -230,4 +230,4 @@ struct AtomicAccess::PlatformOrderedStore<8, RELEASE_X_FENCE>
 };
 #endif // AMD64
 
-#endif // OS_CPU_LINUX_X86_ATOMIC_LINUX_X86_HPP
+#endif // OS_CPU_BSD_X86_ATOMICACCESS_BSD_X86_HPP
